Add types and extract methods to utils
