+++ ui/components/src/TimeSeriesTooltip/tooltip-model.ts
@@ -12,7 +12,7 @@
 // limitations under the License.
 
 import { useEffect, useState } from 'react';
-import { NearbySeriesArray, NearbySeriesInfo } from './nearby-series';
+import { NearbySeriesArray, NearbySeriesInfo } from './types';
 
 export const TOOLTIP_MIN_WIDTH = 375;
 export const TOOLTIP_MAX_WIDTH = 650;
